% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{add_to_logfile}
\alias{add_to_logfile}
\title{Track repository manipulation}
\usage{
add_to_logfile(username, data_level, entry_id, fn_name, fn_input)
}
\arguments{
\item{username}{The name of the user currently logged in.}

\item{data_level}{The current data level.}

\item{entry_id}{The ID of the entry that is manipulated.}

\item{fn_name}{The name of the function with which the repository was manipulated.}

\item{fn_input}{The input of the function with which the repository was manipulated.}
}
\value{
Invisible \code{TRUE}if everything went fine.
}
\description{
Tracks repository manipulation by adding it to the logfile.
}
