% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/argument_dummy.R
\name{argument_dummy}
\alias{argument_dummy}
\title{Argument dummy}
\usage{
argument_dummy()
}
\arguments{
\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}

\item{verbose}{Logical. If \code{TRUE}, informative messages are printed.}
}
\description{
Description dummy for arguments.
}
