% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{assembleLabOrgaDir}
\alias{assembleLabOrgaDir}
\alias{assembleLabOrgaStorageDir}
\title{Obtain configured storage directory}
\usage{
assembleLabOrgaDir()

assembleLabOrgaStorageDir()
}
\value{
Character value.
}
\description{
Retrieves the directory to folder ~/ProgramData and assembles
the required \code{laborga} directories.
}
