% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{checkLaborgaConfiguration}
\alias{checkLaborgaConfiguration}
\title{Checks \code{laborga}-configuration}
\usage{
checkLaborgaConfiguration(in_shiny = FALSE)
}
\arguments{
\item{in_shiny}{Logical.}
}
\value{
\code{TRUE} if everything is fine. \code{FALSE} and warnings in case of problems.
}
\description{
Checks if \code{laborga}-configuration is valid
and returns warnings in case of negative test results.
}
