% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{checkpoint}
\alias{checkpoint}
\title{Shiny feedback messages}
\usage{
checkpoint(
  evaluate = TRUE,
  case_false = NULL,
  error_notifications = list(invalid_password =
    "The password must only contain letters, numbers, !, ?, _ or -.", invalid_project_dir
    = "The storage directory is invalid.", invalid_username =
    "Username must not contain anything else than letters and empty space.",
    no_entry_selected = "Please select a row in the table.", no_project_ids =
    "Please add at least one data entry.", non_matching_passwords =
    "Passwords must be identical.", username_taken = "Username is already taken.",
    username_unknown = "Username is unknown.", 
     zero_tissue_portions =
    "Number of tissue portions to add must not be 0."),
  warning_notifications = list(),
  duration = 4,
  stop_process = TRUE,
  stop_app = FALSE
)
}
\arguments{
\item{evaluate}{A vector of logical tests to be evaluated.}

\item{case_false}{A character string indicating the message to be displayed if one element of
\code{evaluate} turns out to be FALSE. Needs to be in \code{base::names(\code{error/warning_notifiations})}.}

\item{error_notifications}{A named list of character strings.}

\item{warning_notifications}{A named list of character strings.}

\item{duration}{The duration the message is displayed.}

\item{stop_process, stop_app}{Logical. What is supposed to happen if one element of \code{evaluate}
turns out to be FALSE.}
}
\value{
A shiny notification.
}
\description{
Wrapper around \code{shiny::req()} and \code{shiny::showNotification()}.
Prevents application from crashing and displays guiding message about what the user
is supposed to do in order to continue without this message to appear.
}
