% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{configure_lab_orga}
\alias{configure_lab_orga}
\title{Configure \code{LabOrga}}
\usage{
configure_lab_orga()
}
\value{
Invisible \code{TRUE} if successful.
}
\description{
Creates configuration folder via \code{get_dir_configuration()} and
creates the configuration object.
}
