% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{configureStorageDirectory}
\alias{configureStorageDirectory}
\title{Configure \code{laborga} storage directory}
\usage{
configureStorageDirectory(dir)
}
\arguments{
\item{dir}{Character string specifying the storage directory path for \code{laborga}.}
}
\value{
An invisible \code{TRUE} if configuration was succesfull. An invisible \code{FALSE}, if not.
}
\description{
This function allows the user to configure the \code{laborga} storage directory by setting the storage
directory path. It checks if the directory exists and prompts the user to create it if it
doesn't exist. The user is then asked to confirm the directory path before setting it as
an environment variable
}
\details{
The \code{configureLaborga()} function is used to configure the \code{laborga} storage directory. It takes
a directory path as input and performs the following steps:

\itemize{
\item Normalizes the directory path using \code{normalizePath()}.
\item Checks if the directory exists using \code{dir.exists()}.
\item If the directory doesn't exist, prompts the user to create it.
\item If the user confirms to create the directory, it uses \code{dir.create(recursive = TRUE)} to create it.
\item Asks the user to confirm the directory path.
\item If the user confirms, sets the directory path as an environment
variable named LABORGA_STORAGE_DIRECTORY using \code{Sys.setenv()}.
}

If the directory already exists the user is asked to overwrite it. If it does not exist
the user is asked if the directory should be created. If he chooses not to create it, the function is
aborted and the \code{laborga} storage directory is not configured.
}
\examples{
# Configure the `laborga` storage directory
configureLaborga("/path/to/laborga")

}
\seealso{
\code{\link[=configureStorageTables]{configureStorageTables()}}
}
