% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{configureStorageTables}
\alias{configureStorageTables}
\title{Configure \code{laborga} storage tables}
\usage{
configureStorageTables(overwrite = FALSE, in_shiny = FALSE, verbose = TRUE)
}
\arguments{
\item{overwrite}{Logical. Must be set to \code{TRUE} in case of equal file names
in the storage directory.}

\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}

\item{verbose}{Logical. If \code{TRUE}, informative messages are printed.}
}
\value{
An invisible \code{TRUE} if configuration was succesfull. An invisible \code{FALSE}, if not.
}
\description{
This function configures the four tables in which
the information is stored. It requires prior configuration of the
storage directory via \code{configureStorageDirectory()}.
}
\details{
Configures storage tables with \code{laborga} by writing four empty files
in the configured storage directory.

\itemize{
\item \emph{<storage_dir>/tissue_donor_storage_df.RDS}
\item \emph{<storage_dir>/tissue_sample_storage_df.RDS}
\item \emph{<storage_dir>/tissue_portion_storage_df.RDS}
\item \emph{<storage_dir>/raw_data_storage_df.RDS}
}
}
