% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{configureUserTable}
\alias{configureUserTable}
\title{Configures \code{laborga} users table}
\usage{
configureUserTable(overwrite = FALSE, in_shiny = FALSE, verbose = TRUE)
}
\arguments{
\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}

\item{verbose}{Logical. If \code{TRUE}, informative messages are printed.}
}
\value{
An invisible \code{TRUE} if configuration was succesfull. An invisible \code{FALSE}, if not.
}
\description{
Configures \code{laborga} users table
}
