% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{connect_lab_orga}
\alias{connect_lab_orga}
\title{Connect \code{LabOrga} to repository}
\usage{
connect_lab_orga(dir_repo, verbose = TRUE, in_shiny = FALSE)
}
\arguments{
\item{dir_repo}{Character value. The directory to the repository.}
}
\value{
Invisible \code{TRUE} if successful.
}
\description{
Connects \code{LabOrga} to a \code{LabOrga}-repository.
}
