% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{create_repository}
\alias{create_repository}
\title{Set up a \code{LabOrga}-Repository}
\usage{
create_repository(
  dir_repo,
  overwrite = FALSE,
  in_shiny = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dir_repo}{Directory to the repository.}

\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}

\item{verbose}{Logical. If \code{TRUE}, informative messages are printed.}
}
\value{
Invisible \code{TRUE} if successful.
}
\description{
Sets up a \code{LabOrga}-Repository. By creating the files
in \code{repository_files}.
}
\details{
This function creates the following files within \code{dir_repo} making
it a valid \code{LabOrga} repository:

\itemize{
\item \code{dlevel1_tissue_donor.RDS}: A data frame containing tissue donor information.
\item \code{dlevel2_tissue_sample.RDS}: A data frame containing tissue sample information.
\item \code{dlevel3_tissue_portion.RDS}: A data frame containing tissue portion information.
\item \code{dlevel4_raw_data.RDS}: A data frame containing raw data information.
\item \code{logfile.RDS}: A log file storing relevant information and events.
\item \code{setup.RDS}: A list containing setup details.
\item \code{users.RDS}: A data frame containing user information.
}
}
