% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{createLabOrgaDir}
\alias{createLabOrgaDir}
\alias{createLabOrgaStorageDir}
\title{Create \code{laborga} directories}
\usage{
createLabOrgaDir(in_shiny = FALSE)

createLabOrgaStorageDir()
}
\arguments{
\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}
}
\value{
Invisible \code{TRUE} if successful.
}
\description{
Savely creates directories required for \code{laborga} to work.
}
