% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{disconnect_lab_orga}
\alias{disconnect_lab_orga}
\title{Disconnect LabOrga from repository}
\usage{
disconnect_lab_orga(verbose = TRUE, in_shiny = FALSE)
}
\arguments{
\item{verbose}{Logical. If \code{TRUE}, informative messages are printed.}

\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}
}
\value{
Invisble \code{TRUE}
}
\description{
Disconnects \code{LabOrga} from repository.
}
