% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{get_d_level}
\alias{get_d_level}
\alias{get_d_level.data.frame}
\alias{get_d_level.character}
\title{Obtain data level}
\usage{
get_d_level(x, ...)

\method{get_d_level}{data.frame}(x)

\method{get_d_level}{character}(x)
}
\description{
Obtain data level
}
