% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{get_dir_configuration}
\alias{get_dir_configuration}
\title{Obtain dir to configuration folder}
\usage{
get_dir_configuration()
}
\value{
Character value. Depends on the system and the hard drive set up
of the device.
}
\description{
Obtain dir to configuration folder
}
