% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{get_dir_repository}
\alias{get_dir_repository}
\title{Obtain dir to repository}
\usage{
get_dir_repository()
}
\value{
Character value. Depends on the system and the hard drive set up
of the device.
}
\description{
Obtain dir to repository
}
