% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{getLabOrgaStorageDir}
\alias{getLabOrgaStorageDir}
\title{Obtain \code{laborga} storage directory}
\usage{
getLabOrgaStorageDir(in_shiny = FALSE)
}
\arguments{
\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}
}
\value{
Character.
}
\description{
Assembles \code{laborga} storage directory \strong{AND} initiates
configuration if it does not exist via \code{configureStorageDirectory()}.
}
