% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{getStorageDirectory}
\alias{getStorageDirectory}
\title{Obtain configured storage directory}
\usage{
getStorageDirectory()
}
\value{
Character value.
}
\description{
Retrieves what is currently set as the environment
variable \emph{LABORGA_STORAGE_DIRECTORY}.
}
\examples{

 getStorageDirectory()

}
