% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{is_configured}
\alias{is_configured}
\title{Test configuration}
\usage{
is_configured(show_warnings = FALSE, fdb.fn = "warning", in_shiny = FALSE)
}
\arguments{
\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}
}
\value{
Logical value. \code{TRUE} if directory as obtained by \code{get_dir_configuration()}
exists and the \emph{configuration.RDS} file exists within this directory and is a list.
}
\description{
Tests output of \code{configure_lab_orga()}.
}
