% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{is_connected}
\alias{is_connected}
\title{Test connection to repository}
\usage{
is_connected(show_warnings = FALSE, fdb_fn = "warning", in_shiny = FALSE)
}
\arguments{
\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}
}
\value{
Logical value. \code{TRUE} if configuration object in config directory as
obtained by \code{get_config_obj()}contains a directory in slot \verb{$dirs$repo} that leads to a valid \code{LabOrga}
repository as tested by \code{is_repository()}.
}
\description{
Test connection to repository
}
