% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{is_repository}
\alias{is_repository}
\title{Test validity of repository}
\usage{
is_repository(
  dir_repo = get_dir_repository(),
  show_warnings = FALSE,
  in_shiny = FALSE
)
}
\arguments{
\item{dir_repo}{Directory to a supposed repository.}

\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}
}
\value{
Logical value. \code{TRUE}, if repository directory as obtained by
\code{get_dir_repository()} exists, all repository files are found
and all files are valid.
}
\description{
Tests availability and validity of all expected \code{repository_files}
in a supposed repository folder as set up by \code{create_repository()}.
}
