% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{labOrgaDirExists}
\alias{labOrgaDirExists}
\alias{labOrgaStorageDirExists}
\title{Test existence of \code{laborga} directories}
\usage{
labOrgaDirExists(error = FALSE)

labOrgaStorageDirExists(error = FALSE)
}
\arguments{
\item{error}{Logical. If \code{TRUE}, throws an error if test results
are negative.}
}
\value{
Logical value.
}
\description{
Test existence of \code{laborga} directories
}
