% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{load_logfile_df}
\alias{load_logfile_df}
\alias{load_repo_setup}
\alias{load_storage_df}
\alias{load_users_df}
\title{Load repository files}
\usage{
load_logfile_df()

load_repo_setup()

load_storage_df(d_level)

load_users_df()
}
\value{
Depends on the nature of the file. Currently either a list or a data.frame.
}
\description{
Basic loading functions for \code{repository_files}. In some
cases there are additional \verb{get_*()}-functions with more functionalities.
}
