% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{save_config_obj}
\alias{save_config_obj}
\title{Save configuration object}
\usage{
save_config_obj(obj, verbose = TRUE)
}
\description{
Saves configuration object under the directory of slot
\verb{$dirs$config}.
}
