% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{save_logfile_df}
\alias{save_logfile_df}
\alias{save_repo_setup}
\alias{save_storage_df}
\alias{save_users_df}
\title{Save repository files}
\usage{
save_logfile_df(df)

save_repo_setup(setup)

save_storage_df(df)

save_users_df(df)
}
\value{
Invisible \code{TRUE}, if saving was successfull.
}
\description{
Standardized saving of repository files.
}
