% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{setStorageDirectory}
\alias{setStorageDirectory}
\title{Set \code{laborga} storage directory --- probably not needed}
\usage{
setStorageDirectory(dir, verbose = TRUE)
}
\arguments{
\item{dir}{The directory to be set.}
}
\value{
Invisible \code{TRUE}.
}
\description{
Sets the environment variable LABORGA_STORAGE_DIRECTORY.
}
