% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{storageDirectoryExists}
\alias{storageDirectoryExists}
\alias{storageDirectorySet}
\title{Test storage directory validity}
\usage{
storageDirectoryExists(error = FALSE)

storageDirectorySet(error = FALSE)
}
\arguments{
\item{error}{Logical. If \code{TRUE}, throws an error if test results
are negative.}
}
\value{
Logical value.
}
\description{
Test storage directory validity
}
