% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns-config.R
\name{update_repository}
\alias{update_repository}
\title{Update repository}
\usage{
update_repository(in_shiny = FALSE, verbose = TRUE)
}
\arguments{
\item{in_shiny}{Logical. If \code{TRUE}, messages, warnings and errors are returned
within shiny environemnt.}

\item{verbose}{Logical. If \code{TRUE}, informative messages are printed.}
}
\value{
Invisible \code{TRUE} if updating was successful.
}
\description{
Updates repository analogue to the notes of \emph{versions.txt}
}
\section{Side effects}{

Loads files from connected repository, updates them and saves them
under the same direction, overwriting the old ones.
}

