% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fns.R
\name{write_empty_storage_df}
\alias{write_empty_storage_df}
\title{Write empty storage data.frames}
\usage{
write_empty_storage_df(
  d_level,
  dir,
  overwrite = FALSE,
  verbose = TRUE,
  in_shiny = FALSE
)
}
\description{
Write empty storage data.frames
}
\seealso{
\code{\link[=save_storage_df]{save_storage_df()}}, \code{\link[=load_storage_df]{load_storage_df()}}
}
